Fix database selection on Github workflow
To run the specs against a different CI we need to set the DB
environment variable.
We also have to make sure the credentials match the db/database.yml

The database configuration requires stringified keys for Rails 6.1.

The branch for the latest Rails version has been changed from master to
main.